XP-1002 Rename Module exceptions
