+++ modules/core-api/src/main/java/com/enonic/xp/app/ApplicationNotFoundException.java
@@ -1,15 +1,14 @@
-package com.enonic.xp.module;
+package com.enonic.xp.app;
 
 import com.google.common.annotations.Beta;
 
-import com.enonic.xp.app.ApplicationKey;
 import com.enonic.xp.exception.NotFoundException;
 
 @Beta
-public final class ModuleNotFoundException
+public final class ApplicationNotFoundException
     extends NotFoundException
 {
-    public ModuleNotFoundException( final ApplicationKey applicationKey )
+    public ApplicationNotFoundException( final ApplicationKey applicationKey )
     {
         super( "Module [{0}] was not found", applicationKey );
     }
