+++ modules/core-api/src/main/java/com/enonic/xp/app/ApplicationService.java
@@ -2,13 +2,11 @@ package com.enonic.xp.app;
 
 import com.google.common.annotations.Beta;
 
-import com.enonic.xp.module.ModuleNotFoundException;
-
 @Beta
 public interface ApplicationService
 {
     Application getModule( ApplicationKey key )
-        throws ModuleNotFoundException;
+        throws ApplicationNotFoundException;
 
     Applications getModules( ApplicationKeys keys );
